code: updated Nuget references
fix: added code to  AddToPackageCommand.GetProjectPath and callers to prevent an error when a user right-clicks on a file which is not part of a project (solution files)

feature: Nucleus.DeveloperTools.MSBuild. added a check to warn if the casing of files in package.xml does not match file system casing.  Added checking for project content files missing from package.xml, and for files in package.xml that are not project content files.

feature: improved formatting of package.xml after AddToPackageCommand execution.